% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pull_sql_data.R
\name{pull_sql_data}
\alias{pull_sql_data}
\title{pull_sql_data}
\usage{
pull_sql_data(
  source,
  eAP_product_id,
  product_num,
  last_date,
  type = c("downloads", "uploads")
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
